+++ functions/modules/common/classes/Entity.js
